Version tasks and images
In a real deployment, we want to have immutable tasks: each version of a
task has a unique name, and the task is never updated. The task points
to an image with a matching version, so that the task is not implicitly
updated when the referenced images change.

In local development and testing though, we do not care about the
version. The new approach avoids that developers have to specify a
version in the tests.

As a consequence, the image of each task step is not parameterized
anymore. This is desired as well: if the official ClusterTask is used,
we want to be sure that it actually executed as intended. If users were
able to override the image, they could modify ClusterTasks any way they
want really.

Closes #73.
Closes #74.