Drop concept of phases and finish environment/version support
Both things are close together in code so it was easier to do both
together.

Instead of phases, ods.yml can now specify pipeline tasks under `tasks` and
`finally`. Closes #91.

The pipeline tasks now fully support `environment` and `version`
parameters. However, setting of a proper environment in the interceptor
is not yet supported. For that, we still need to implement support for
`branchToEnvironmentMapping` and externa API requests directly selecting
an environment. Therefore this does not close #60 just yet.

Further, this commit cleans up a few things noticed during the
refactoring.