Remove whitespace from Go templates
Otherwise, unmarshalling and marshalling breaks long lines. This is an
issue e.g. for the `image` field of container definitions - if the
`image` field is made up of Helm values, then those must not be broken
up on whitespace or an error is caused. For consistency in the
templates, all Go templates have been aligned.

The issue was discovered during the implementation of the sidecar
feature, which parses tasks, modifies them, and then writes the "sidecar
task" to a new file.