Make task prefix customizable
While the default is still `ods`, it is now possible to use a
different prefix. That way, an installation of ODS pipeline can be
installed centrally without communicating to cluster users that it is
part of the "official" ODS core installation. For example, one could use
`experimental` as a prefix to indicate that this installation is not (yet)
blessed by the platform owners.

Closes #289