Fix Helm invocation
The quotes determine what makes up one entry in the array. So e.g.
"secrets upgrade" is seen as one entry, not as two. This leads to:

```
Error: unknown command "secrets upgrade" for "helm"
```

The regression was introduced in #322. ShellCheck is nice, but rule
https://github.com/koalaman/shellcheck/wiki/SC2086 in particular might
cause more problems than it solves ... ?!