Adjust helm chart to install published images by default
To be discussed: Until the v0.3.0 release is done, I'd propose to use latest as the image tag. We cannot easily backport to v0.2.0 since several image builds in v0.2.0 state are failing meanwhile, a component (`webhook-interceptor`) was renamed, we don't have release branches (yet) and `ods-pipeline` in general is still pre-1.0.0.