Refactor pipeline manager
Instead of one huge server struct having lots of responsibilities, split
into multiple pieces:

* receiving Bitbucket webhooks
* scheduling pipeline runs
* pruning pipeline runs
* watching queued pipeline runs

This makes it easier to test the functionality in isolation, and this PR
already takes advantage of this. Further, this PR adds more tests for parts of
the pipeline manager which were only tested on a higher level (such as
Bitbucket webhook validation).

As a nice side effect of the refactoring, the mutex is not needed
any longer as the channels synchronize access.

Finally, note that this commit moves the endpoint of the webhook receiver to
`/bitbucket`, as a consequence every webhook configuration in Bitbucket needs
to be updated. The change was made to allow other endpoints easier in
the future.

Closes #491.