Do not add chart labels to tasks
The labels were added recently in #550.

However, I was unaware that Tekton propagates the labels of tasks to the
pods: https://tekton.dev/docs/pipelines/labels/#label-propagation.

This propagation has the unwanted effect that the `ods-pipeline`
deployment now considers all pods that are created from pipelines as
part of the deployment.

Unfortunately there seems to be no way to have the labels for the tasks,
but not propagated to the pods.

An alternative solution would have been to change the selector in the
deployment, but that would not change that the pods are labelled as
being "managed by Helm", which isn't correct so instead we just do not
add labels to the tasks, even though that also does not look right.
However technically it should be fine (and #550 needs the labels only
for the build config resources).

Fixes #555.