Massage confusing Helm diff output
Originally implemented in #268, however the test wasn't bullet-proof, so
it did not fail when #469 was merged. Since then, the log output
replacement didn't work. Further, the detection did not work when debug
mode was enabled as that produced more output than expected.

The new approach first detects if there is a diff, then tries to remove
all offending output.

Closes #563.