Refactor deployment task logic
Instead of having all the logic written out in main.go, the logic is now
broken up into smaller chunks, using a functional options style. The
motivation behind this is to facilitate #589.

Along with the main change, a few minor things were changed as well such
as cleanup of error handling and messages, logging, etc.