Test that downloaded binary works
It seems that Aqua is serving an empty binary sometimes. Because the
status code is 200, and the resulting file is executable, there is no
error until the scanner is invoked. The code change prevents this by
testing if the output of the version command is empty. It's a very basic
check but should catch the weird behaviour we have now observed a few
times.

See #638.