Do not error when no matching pipeline can be found
Returning an error does not seem appropriate because there can be
situations in which no pipeline should match which are "expected". For
example, one might configure to act on PR comments of "/deploy", but for
all other PR comment (events), nothing should happen.

The log will still report that there was no matching pipeline so that
users can figure out what is going on.

The issue was introduced in #685. Since that has not been released yet,
the bugfix will not be recorded in the changelog.