Allow to execute diff only, without upgrade.
This necessitates that the image promotion happens after the diff, which
is something we wanted to do anyway.

The new feature plays nicely with the newly introduced trigger-based
params as those allow to set diff-only for certain cases, e.g. only when
a PR comment is made.

Closes #589.
Solves #622 partially. It runs the diff but does not post to Bitbucket
yet. Not perfect, but already good enough to provide value.