Consult artifact repository in artifact check
Artifacts may be downloaded in the start task from repository A, but
should be uploaded to repository B in the finish task. In this case,
Contains erroneously returned true before this change. Contains should
only return true if the artifacts were downloaded from the same
repository earlier. When the repositories differ, the task must fetch a
manifest from the target repository before checking the manifest.