Fix Go module name
For some reason the Go module was 'github.com/opendevstack/pipeline'
while the repository is named 'github.com/opendevstack/ods-pipeline'.
This means that importing packages from this module was not possible
before. While this isn't a typical thing to do now, it will be once we
move the tasks out to their own repositories. I have a PoC of this
locally which requires the change here ...