Do not require bitbucketUrl
While helpful in general, this breaks tests of tasks that do not use
Bitbucket. Going forward we should find a way to require the Url but set
it e.g. to https://example.com when tests do not require Bitbucket. For
now, this is just a quick fix to unblock tests from upgrading.